requirements.txt
python-dotenv
requests
uvicorn
fastapi
python-dotenv